Update mypy.yml
Running `mypy` in the worker always fails (the command, not the type checks, but those fail too). I believe that it's because of how mypy processes imports.
I was able to fix it locally by using the `--explicit-package-bases` option. But to be able to pass that in the CI we need to expand this a little bit.